Implemented both cylinder caps'
