+++ test_rt_files/cy.rt
@@ -3,6 +3,7 @@ C  0,0,0 0,0,1 120
 L 10,10,10 0.6 255,255,255
 pl 0,-200,0 0,1,0 102,255,178
 
-sp -12,20,20 4 255,100,255
 
-cy 0,0,50 0,0,1 10 20 255,100,255
+sp -10,0,20 10 255,153,204
+
+cy 10,0,25 0,0,1 10 10 255,100,255
